Removed 2 unnecessary stubbings in AmazonSQSExtendedClientTest.java
